requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
rapidfuzz>=3.0.0
pillow>=10.0.0
python-Levenshtein>=0.20.0

# Computer Vision and AI - Cloud deployment friendly versions
opencv-python-headless>=4.8.0
torch>=2.0.0,<2.3.0
torchvision>=0.15.0,<0.18.0
transformers>=4.30.0,<4.40.0
scikit-learn>=1.3.0
numpy>=1.24.0,<1.26.0
datasets>=2.14.0
scipy>=1.11.0
